break variable entries into their own file for model/parserTypes cycle
