+++ src/vs/workbench/contrib/chat/common/chatVariables.ts
@@ -9,8 +9,9 @@ import { URI } from '../../../../base/common/uri.js';
 import { IRange } from '../../../../editor/common/core/range.js';
 import { Location } from '../../../../editor/common/languages.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
-import { IChatModel, IDiagnosticVariableEntryFilterData } from './chatModel.js';
+import { IChatModel } from './chatModel.js';
 import { IChatContentReference, IChatProgressMessage } from './chatService.js';
+import { IDiagnosticVariableEntryFilterData } from './chatVariableEntries.js';
 import { IToolData, ToolSet } from './languageModelToolsService.js';
 
 export interface IChatVariableData {
